+++ CHANGELOG.md
@@ -4,7 +4,9 @@ Yii 2 File DB extension Change Log
 1.0.3 under development
 -----------------------
 
+- Enh: `QueryProcessor::filterHashCondition()` allows to specify filter value as `\Closure` instance (klimov-paul)
 - Enh #5: `QueryProcessor::filterInCondition()` advanced allowing comparison against array-type columns (klimov-paul)
+- Enh #6: Added `QueryProcessor::filterCallbackCondition()` allowing to specify PHP callback as filter (klimov-paul)
 
 
 1.0.2, June 28, 2016